log free mem only if it changes

+++ PlantMonitor/Util.cpp
@@ -27,8 +27,8 @@ uint32_t util_millis() {
 	return cycleMilis;
 }
 
-uint32_t util_freeRam() {
+uint16_t util_freeRam() {
 	extern int __heap_start, *__brkval;
 	int v;
-	return (int) &v - (__brkval == 0 ? (int) &__heap_start : (int) __brkval);
+	return (uint16_t) &v - (__brkval == 0 ? (int) &__heap_start : (int) __brkval);
 }
